requirement.txt
arrow==1.4.0
asgiref==3.11.0
binaryornot==0.4.4
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cookiecutter==2.6.0
diff-match-patch==20241021
Django==6.0
django-import-export==4.3.14
django-tailwind==4.4.2
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
six==1.17.0
sqlparse==0.5.4
tablib==3.9.0
text-unidecode==1.3
tzdata==2025.3
urllib3==2.6.2
